Consolidated files and changed request syntax
