--- src/model/armor.ts
@@ -1,20 +1,0 @@
-export class Armor{
-    public id!: number;
-    public slug!: string;
-    public name!: string;
-    public type!: string;
-    public rank!: string;
-    public rarity!: number;
-    public defense!: object;
-    public resis!: object;
-    public slots!: object[];
-    public attrib!: object;
-    public skills!: object[];
-    public armorSet!: object;
-    public assets!: object;
-    public crafting!: object;
-
-    constructor(){
-        
-    }
-}
