--- src/model/armorsets.ts
@@ -1,11 +1,0 @@
-export class Armorsets{
-    public id!: number;
-    public name!: string;
-    public rank!: string;
-    public pieces!: object[];
-    public bonus!: object;
-
-    constructor(){
-        
-    }
-}
