--- src/model/weapons.ts
@@ -1,20 +1,0 @@
-export class Weapons{
-    public id!: number;
-    public name!: string;
-    public type!: string;
-    public rarity!: number;
-    public att!: object;
-    public elderseal!: any;
-    public attrib!: object;
-    public damageType!: string;
-    public durability!: object[];
-    public slots!: object[];
-    public crafting!: object;
-    public craftingMat!: object[];
-    public upgradeMat!: object[];
-    public assets!: object;
-
-    constructor(){
-        
-    }
-}
