+++ src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const mhwURL: string = "https://mhw-db.com";
